+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.2.15
+- [iOS] PUSH notifications handling was added to the example app.
+- [Android] 'optOutOfTracking' and 'collectDeviceIdentifiers' are introduced
+- [Android] **BREAKING** Method `disableAdTracking` is removed.
+
+- Dependencies of Native SDK's were updated to:
+  - [Android] 1.8.0
+
 ## 2.2.14
 - [Android] Fixes bug with AddAttribution for AppsFlyer provider.